complain if help is given an unknown command
This requires that I stop using the if-else chain, and instead add a
second output that tracks whether a command match OR a help match
occurred. This might seem less pleasing, but really it makes the code
more order-independent anyway, so it's probably nicer overall as well.

Also in making the match_keyword procedures more complicated, I decided
to make match_or_explain_keyword call into match_keyword, almost like a
wrapper. The redundant help print code is fine, because soon we will be
printing different messages in each of those cases, depending on whether
we are in the general help message, or in some command-specific help
information.