add simple and detailed match procedures
Simple is like before, but specifically requires that no arguments are
given, or else it will display an error.

Detailed is like before, but with a more detailed message when help is
being run on that command specifically.

Also help messages now need newlines to be added explicitly, since in a
text-adventure you might actually be using `look` instead of `help`, and
in that case you would want the option to list all the objects in the
space as a single sentence, without newlines in between.